Separate SortableTalkList into another file.
